create update and delete actions on article, profile, projects forms

+++ app/Http/Controllers/API/ArticleController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Article;
 use App;
+use Illuminate\Support\Facades\DB;
 use Validator;
 use App\Http\Resources\Article as ArticleResource;
 
@@ -95,10 +96,11 @@ class ArticleController extends BaseController
         $input = $request->all();
         $aid =$request->user_id;
         $usid = auth()->user()->id;
+        //return $article;
 
         if($aid!=$usid)
         {
-            return $this->sendError('Sorry, edit your own articles.');
+            return $this->sendError('Sorry, edit your own article.');
         }
 
         $validator = Validator::make($input, [
@@ -139,12 +141,25 @@ class ArticleController extends BaseController
     public function destroy($id)
     {
         //$article->delete();
+       // global $aid;
 
-        Article::destroy($id);
+        $aid = DB::table('articles')->where('id', $id)->pluck('user_id')->first();
 
-       // return $this->sendResponse([], 'Article deleted successfully.');
+        $usid = auth()->user()->id;
+        //return $usid;
 
-        return response()->json("ok");
+        if($aid !== $usid)
+        {
+            return $this->sendError('Sorry, you can delete your own article.');
+        }
+
+        else {
+            Article::destroy($id);
+
+            // return $this->sendResponse([], 'Article deleted successfully.');
+
+            return response()->json("ok");
+        }
         //return view('articles.show', ['articles' => $article]);
     }
