+++ app/Http/Controllers/API/ProfileController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Profile;
 use App;
+use Illuminate\Support\Facades\DB;
 use Validator;
 use App\Http\Resources\Profile as ProfileResource;
 
@@ -178,6 +179,15 @@ class ProfileController extends BaseController
     public function delete($id)
     {
         //$profile->delete();
+
+        $pid = DB::table('profiles')->where('id', $id)->pluck('user_id')->first();
+
+        $usid = auth()->user()->id;
+        if($pid !== $usid)
+        {
+            return $this->sendError('Sorry, you can delete your own profile.');
+        }
+
         Profile::destroy($id);
 
         //return $this->sendResponse([], 'Profile deleted successfully.');
