+++ app/Http/Controllers/API/ProjectController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Project;
 use App;
+use Illuminate\Support\Facades\DB;
 use Validator;
 use App\Http\Resources\Project as ProjectResource;
 
@@ -137,6 +138,14 @@ class ProjectController extends BaseController
     public function destroy($id)
     {
         //$project->delete();
+
+        $pid = DB::table('projects')->where('id', $id)->pluck('user_id')->first();
+
+        $usid = auth()->user()->id;
+        if($pid !== $usid)
+        {
+            return $this->sendError('Sorry, you can delete your own project.');
+        }
 
         Project::destroy($id);
